refactor(repo C14a): extract retrieval primitives to engine/repository/retrieval/
Move expandQuery, rerank, reciprocalRankFusion + buildRrfTrace + RankedListMeta,
and the document-lookup family (findDocument, getDocumentBody, findDocuments,
getStatus, extractSnippet, addLineNumbers, withTimeout, hasDocumentIngestTable)
into engine/repository/retrieval/. The orchestrators hybridQuery,
vectorSearchQuery, structuredSearch remain in engine/repository.ts and now
import their dependencies from the new retrieval modules.

Build: clean. Tests: 970/970 pass.
